requirements.txt
fastapi==0.111.0
pydantic==1.10.7
python-ldap==3.4.4
requests==2.28.2
